pyproject.toml
# if lock file, poetry will install all those but with
# exact versions from lock file
# otherwise it updates them.
# lock file has to be commited for exact reprod.

[tool.poetry]
authors = ["Santiago Miranda"]
description = "New Chemical Variational Autoencoder."
license = "Apache 2.0"
name = "cvae"
readme = "README.md"
packages = [{ include = "chemvae" }]
package-mode = true
version="0.0.0"


[project.urls]
changelog = "https://github.com/ghsanti/cvae/blob/master/CHANGELOG.md"
documentation = "https://cvae.readthedocs.io"
homepage = "https://cvae.readthedocs.io"
issues = "https://github.com/ghsanti/cvae/issues"
repository = "http://github.com/ghsanti/cvae.git"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
my-script = "chemvae.log_revision:start"

[tool.poetry.dependencies]
keras = "3.3.3"
numpy = "1.26.*"
pandas = "^2.2.2"
python = "~3.12.4" # install from conda, use --no-deps (still testing this!)
pyyaml = "^6.0.1"
rdkit = "^2023.9.6"
tensorflow = {extras = ["and-cuda"], version = "^2.16.1"}

pydot = "^2.0.0"
matplotlib = "^3.9.0"
[tool.poetry.group.dev.dependencies]
pandas-stubs = "^2.2.2.240603"
pre-commit = "^3.7.1"
ruff = "^0.4.9"
types-pyyaml = "^6.0.12.20240311"
types-tensorflow = "^2.16.0.20240618"
pyright = "^1.1.368"

sphinx = "^7.3.7"
pdoc3 = "^0.10.0"
python-semantic-release = "^9.8.3"
mkdocs = "^1.6.0"
mkdocstrings = {extras = ["python"], version = "^0.25.1"}
[tool.ruff]
# remove excludes, use include is shorter.
include = ["pyproject.toml", "chemvae/**/*.py"]
exclude = ["**/*.ipynb"]

indent-width = 4
line-length = 88
target-version = "py312"
show-fixes = true

[lint]
fixable = ["ALL"]
ignore = []
# select = ["D", "I", "F", "E", "N", "W", "TID", "SIM", "ARG", "PTH", "ERA", "B"]
select = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["chemvae"]
forced-separate = ["tests"]
relative-imports-order = "closest-to-furthest"
section-order = [
  "future",
  "standard-library",
  "third-party",
  "first-party",
  "tests",
  "local-folder",
]
sections = { "tests" = ["tests"] }

[format]
docstring-code-format = true
docstring-code-line-length = "dynamic"
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]

[tool.pyright]
include=["chemvae/**"]
exclude=["**/__init__.py", "**/__pycache__"]
verboseOutput=true
pythonVersion="3.11"
typeCheckingMode="standard"
reportMissingImports= true
reportMissingTypeStubs= false
[tool.semantic_release]
assets = []
build_command_env = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
template_dir = "templates"
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = []

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = true

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
default_bump_level = 0

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
